Moving GraphData to routing_common.
