+++ routing_common/CMakeLists.txt
@@ -1,5 +1,9 @@
 project(routing_common)
 
+include_directories(
+  ${OMIM_ROOT}/3party/jansson/src
+)
+
 set(
   SRC
   bicycle_model.cpp
@@ -9,6 +13,8 @@ set(
   num_mwm_id.hpp
   pedestrian_model.cpp
   pedestrian_model.hpp
+  transit_graph_data.cpp
+  transit_graph_data.hpp
   transit_serdes.hpp
   transit_speed_limits.hpp
   transit_types.cpp
